+++ pycheops/starproperties.py
@@ -35,7 +35,7 @@ from .core import load_config
 from pathlib import Path
 from os.path import getmtime
 from time import localtime, mktime
-from uncertainties import ufloat
+from uncertainties import ufloat, UFloat
 from .ld import stagger_power2_interpolator
 from numpy.random import normal
 
@@ -47,7 +47,9 @@ class StarProperties(object):
     """
 
     def __init__(self, identifier, force_download=False, 
-            match_arcsec=2, configFile=None, verbose=True):
+            match_arcsec=5, configFile=None,
+            teff=None, logg=None, metal=None, 
+            verbose=True):
 
         self.identifier = identifier
         coords = SkyCoord.from_name(identifier)
@@ -91,18 +93,45 @@ class StarProperties(object):
             raise ValueError('No matching star in SWEET-Cat')
 
         entry = sweetCat[idx]
+        ld_data_missing = False
         try:
             self.teff = ufloat(float(entry['teff']),float(entry['e_teff']))
+            self.teff_note = "SWEET-Cat"
         except:
             self.teff = None
+            ld_data_missing = True
         try:
             self.logg = ufloat(float(entry['logg']),float(entry['e_logg']))
+            self.logg_note = "SWEET-Cat"
         except:
             self.logg = None
+            ld_data_missing = True
         try:
             self.metal = ufloat(float(entry['metal']),float(entry['e_metal']))
+            self.metal_note = "SWEET-Cat"
         except:
             self.metal = None
+            ld_data_missing = True
+
+        # User defined values
+        if teff is not None:
+           if  isinstance(teff, UFloat):
+               self.teff = teff
+               self.teff_note = "User"
+           else:
+               raise ValueError("teff keyword is not ufloat")
+        if logg is not None:
+           if  isinstance(logg, UFloat):
+               self.logg = logg
+               self.logg_note = "User"
+           else:
+               raise ValueError("logg keyword is not ufloat")
+        if metal is not None:
+           if  isinstance(metal, UFloat):
+               self.metal = metal
+               self.metal_note = "User"
+           else:
+               raise ValueError("metal keyword is not ufloat")
 
         # log rho from log g using method of Moya et al.
         # (2018ApJS..237...21M). Accuracy is 4.4%
@@ -112,21 +141,21 @@ class StarProperties(object):
                 logrho = -7.352 + 1.6580*self.logg
                 self.logrho = ufloat(logrho.n, np.hypot(logrho.s, 0.044))
 
-        power2 = stagger_power2_interpolator()
-        _,_,h_1,h_2 = power2(self.teff.n,self.logg.n,self.metal.n)
-        if np.isnan(h_1):
-            self.h_1 = None
-            self.h_2 = None
-        else:
-            Xteff = normal(self.teff.n, self.teff.s, 256)
-            Xlogg = normal(self.logg.n, self.logg.s, 256)
-            Xmetal = normal(self.metal.n, self.metal.s, 256)
-            X = power2(Xteff,Xlogg,Xmetal)
-            # Additinal error derived in Maxted, 2019
-            e_h_1 = np.hypot(0.01,np.sqrt(np.nanmean((X[:,2]-h_1)**2)))
-            e_h_2 = np.hypot(0.05,np.sqrt(np.nanmean((X[:,3]-h_2)**2)))
-            self.h_1 = ufloat(round(h_1,3),round(e_h_1,3))
-            self.h_2 = ufloat(round(h_2,3),round(e_h_2,3))
+        self.h_1 = None
+        self.h_2 = None
+        if not ld_data_missing:
+            power2 = stagger_power2_interpolator()
+            _,_,h_1,h_2 = power2(self.teff.n,self.logg.n,self.metal.n)
+            if not np.isnan(h_1):
+                Xteff = normal(self.teff.n, self.teff.s, 256)
+                Xlogg = normal(self.logg.n, self.logg.s, 256)
+                Xmetal = normal(self.metal.n, self.metal.s, 256)
+                X = power2(Xteff,Xlogg,Xmetal)
+                # Additinal error derived in Maxted, 2019
+                e_h_1 = np.hypot(0.01,np.sqrt(np.nanmean((X[:,2]-h_1)**2)))
+                e_h_2 = np.hypot(0.05,np.sqrt(np.nanmean((X[:,3]-h_2)**2)))
+                self.h_1 = ufloat(round(h_1,3),round(e_h_1,3))
+                self.h_2 = ufloat(round(h_2,3),round(e_h_2,3))
 
 
 
@@ -134,14 +163,14 @@ class StarProperties(object):
         s =  'Identifier : {}\n'.format(self.identifier)
         s +=  'Coordinates: {} {}\n'.format(self.ra, self.dec)
         if self.teff is not None:
-            s += 'T_eff : {:5.0f} +/- {:0.0f} K\n'.format(
-                    self.teff.n, self.teff.s)
+            s += 'T_eff : {:5.0f} +/- {:0.0f} K    [{}]\n'.format(
+                    self.teff.n, self.teff.s,self.teff_note)
         if self.logg is not None:
-            s += 'log g : {:5.2f} +/- {:0.2f} [cgs]\n'.format(
-                    self.logg.n, self.logg.s)
+            s += 'log g : {:5.2f} +/- {:0.2f}    [{}]\n'.format(
+                    self.logg.n, self.logg.s, self.logg_note)
         if self.metal is not None:
-            s += '[M/H] : {:+5.2f} +/- {:0.2f}\n'.format(
-                    self.metal.n, self.metal.s)
+            s += '[M/H] : {:+5.2f} +/- {:0.2f}    [{}]\n'.format(
+                    self.metal.n, self.metal.s, self.metal_note)
         if self.logrho is not None:
             s += 'log rho : {:5.2f} +/- {:0.2f} [solar]\n'.format(
                     self.logrho.n, self.logrho.s)